requirements.txt
numpy
scikit-fuzzy
matplotlib
paho.mqtt
matplotlib